requirements.txt
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
nltk>=3.9
streamlit>=1.36.0
pyyaml>=6.0.1
tqdm>=4.66.0
keybert>=0.5.1
joblib>=1.3.0
